+++ src/app/components/dynamic-module-loader/data/terra-split-view-component.interface.ts
@@ -2,6 +2,7 @@ import { TerraMultiSplitViewInterface } from '../../split-view/multi/interfaces/
 
 /**
  * @author pweyrich
+ * @deprecated
  */
 export class TerraSplitViewComponentInterface
 {