+++ src/app/components/example-collection.ts
@@ -47,6 +47,7 @@ import { TerraInfoComponentExample } from './info/example/terra-info.component.e
 import { TerraFormsExampleComponent } from './forms/terra-forms.example';
 import { RadioGroupComponentExample } from './forms/input/radio/example/radio-group.component.example';
 import { TerraDataTableTemplateComponentExample } from './tables/data-table/example/terra-data-table-template.component.example';
+import { FilterComponentExample } from './filter/example/filter.component.example';
 
 export const examples:Array<Type<any>> = [
     TerraButtonComponentExample,
@@ -96,5 +97,6 @@ export const examples:Array<Type<any>> = [
     TerraInfoComponentExample,
     TerraFormsExampleComponent,
     RadioGroupComponentExample,
-    TerraDataTableTemplateComponentExample
+    TerraDataTableTemplateComponentExample,
+    FilterComponentExample
 ];